requirements.txt
bcrypt
pyjwt
fastapi
pymongo = python -m pip install "pymongo[srv]"
cloudinary